Add basic mobile inference runtime
